Changed title of class Map to not interfere with possible use of an actual Map.
Created subclasses for Blocks.
These can be entered as a Block for the pairing with an id, so the "block-type" can be more easily called.

Written intend to use Map or Pair to bundle an integer (id), and block (which can be a specific subclass).

Introduced structure to generate 2d int arrays for referencing different types of objects.